refactor(build): keep install deps in one place
Problem: If we list install deps (that are not git submodules, e.g.,
numpy or pandas) in pyproject.toml, we will have to list them in
requirements.txt as well.  And we need requirements.txt because we need
to install the git submodules in editable form, which we do in
requirements.txt.

Solution: Use dynamic dependency approach of setuptools.  It will read
from requirements.txt file.  The setuptools extension will need to parse
the "-e /path/to/submodule" lines and generate the actual package name
and version from it.